null photos fixed, album title added

+++ client/js/photos.js
@@ -1,33 +1,56 @@
 document.addEventListener("DOMContentLoaded", () => {
     loadPhotos();
+    loadAlbumInfo();
     const form = document.querySelector("#formy-mcformface")
     form.addEventListener("submit", addPhoto)
 });
 
+const loadAlbumInfo = async () => {
+    let params = (new URL(document.location)).searchParams;
+    let album = params.get("album");
+    let response = await axios.get(`http://localhost:11000/albums/${parseInt(album)}`);
+    let albumInfo = response.data.payload
+    console.log(albumInfo)
+    for (let singleAlbum of albumInfo) {
+        let params = (new URL(document.location)).searchParams;
+        let album = params.get("album");
+        if (singleAlbum.album_id === parseInt(album)) {
+            let title = document.querySelector("#album-title")
+            title.innerText = singleAlbum.album_title
+        }
+    }
+}
 
 const loadPhotos = async () => {
     let params = (new URL(document.location)).searchParams;
     let album = params.get("album");
     let response = await axios.get(`http://localhost:11000/photos/albums/${parseInt(album)}`);
     let existingPhotos = response.data.body
-    for (let photo of existingPhotos) {
+    if (existingPhotos) {
+        for (let photo of existingPhotos) {
+            let mainDiv = document.querySelector("#photo-container")
+            let newDiv = document.createElement("div")
+            newDiv.id = photo.photo_id
+            newDiv.className = "picture-frame"
+            let newPic = document.createElement("img")
+            newPic.src = photo.photo_url
+            newPic.className = "user-photo"
+            let picTitle = document.createElement("p")
+            picTitle.innerText = photo.title
+            let deleteBtn = document.createElement("button")
+            deleteBtn.innerText = "DELETE"
+            deleteBtn.className = "button"
+            newDiv.append(picTitle, newPic, deleteBtn)
+            mainDiv.append(newDiv)
+            deleteBtn.addEventListener("click", () => {
+                deletePhoto(newDiv.id);
+            }) 
+        }
+    } else {
         let mainDiv = document.querySelector("#photo-container")
-        let newDiv = document.createElement("div")
-        newDiv.id = photo.photo_id
-        newDiv.className = "picture-frame"
-        let newPic = document.createElement("img")
-        newPic.src = photo.photo_url
-        newPic.className = "user-photo"
-        let picTitle = document.createElement("p")
-        picTitle.innerText = photo.title
-        let deleteBtn = document.createElement("button")
-        deleteBtn.innerText = "DELETE"
-        deleteBtn.className = "button"
-        newDiv.append(picTitle, newPic, deleteBtn)
-        mainDiv.append(newDiv)
-        deleteBtn.addEventListener("click", () => {
-            deletePhoto(newDiv.id);
-        }) 
+        let alert = document.createElement("p")
+        alert.innerText = "Album empty! Please add a photo to this album by using the input above!"
+        mainDiv.append(alert)
     }
 }
